🛠️ Refactor: Split API handler into focused files
